GimmePeers Blah Query Value Update (#338)
* GimmePeers Blah Query Value Update

GP recently released an updated IMDB ID search option for queries. This change was added because queries against "blah=3" were turning up inconsistent results because not all "title" or "NFO" contained the IMDb. The default "blah" value in the gimmepeers.yaml file is 3, but it should now be changed to 4.

* GP Blah Query Value Update

Revised the "blah" line to only set 4 when the IMDb id is queried, but continue to use "blah=3" for Genre.

* GP Blah Query Value Update

Revised to match the Jackett update for the same "blah" query.

https://github.com/Jackett/Jackett/commit/927ffdff390aac2586d9890a4b3ee5f82d73789e